removed redundant file added a loading spinner and finished the device id selection
