+++ client/src/components/LandingPage.tsx
@@ -1,23 +1,36 @@
 import { useNavigate} from "react-router-dom"
-import { useState } from "react"
+import { useState, useEffect } from "react"
 import { connect } from 'react-redux'
+import { useDispatch, useSelector } from 'react-redux';
 import useFirebase from '../hooks/useFirebase';
-import { useDispatch } from 'react-redux';
 
-import  { setIsHostTrue, setIsHostFalse, setHasJoinedFalse } from '../store/slices/videoCallSlice';
+
+import  { setIsHostTrue, setIsHostFalse, setHasJoinedFalse, setRefreshedFalse } from '../store/slices/videoCallSlice';
 
 const LandingPage = () => {
   const dispatch = useDispatch()
   const navigation = useNavigate();
   const [joinCode, setJoinCode] = useState('645a205090d5f0e0b2b99689');
+  const refresh = useSelector((state: any) => state.videoCall.hasRefreshed)
 
-  // const { db } = useFirebase()
+  useEffect(() => {
+
+    dispatch(setRefreshedFalse())
+  }, [])
+
+
+
+
+
+
+
 
   const handleCreateCall = () => {
     dispatch(setIsHostTrue())
     dispatch(setHasJoinedFalse())
 
     //TODO generate room id from firebase (or uuid? still unclear)
+    // const { db } = useFirebase() //TODO: use this to generate room id?
     //!make post request to localhost:3001/room
     // wait for response
     // fail gracefully if needed
